+++ node/npm/glide/index.ts
@@ -126,6 +126,7 @@ function initialize() {
         VectorFieldAttributesFlat,
         VectorFieldAttributesHnsw,
         FtCreateOptions,
+        FtSearchOptions,
         FtInfoReturnType,
         FtAggregateOptions,
         FtAggregateLimit,
@@ -135,9 +136,11 @@ function initialize() {
         FtAggregateSortBy,
         FtAggregateSortProperty,
         FtAggregateApply,
+        FtSearchReturnType,
         GlideRecord,
         GlideString,
         JsonGetOptions,
+        JsonArrPopOptions,
         SortedSetDataType,
         StreamEntryDataType,
         HashDataType,
@@ -221,6 +224,8 @@ function initialize() {
         ReturnTypeMap,
         ClusterResponse,
         ReturnTypeAttribute,
+        ReturnTypeJson,
+        UniversalReturnTypeJson,
     } = nativeBinding;
 
     module.exports = {
@@ -253,6 +258,7 @@ function initialize() {
         VectorFieldAttributesFlat,
         VectorFieldAttributesHnsw,
         FtCreateOptions,
+        FtSearchOptions,
         FtInfoReturnType,
         FtAggregateOptions,
         FtAggregateLimit,
@@ -262,10 +268,12 @@ function initialize() {
         FtAggregateSortBy,
         FtAggregateSortProperty,
         FtAggregateApply,
+        FtSearchReturnType,
         GlideRecord,
         GlideJson,
         GlideString,
         JsonGetOptions,
+        JsonArrPopOptions,
         SortedSetDataType,
         StreamEntryDataType,
         HashDataType,
@@ -363,6 +371,8 @@ function initialize() {
         ReturnTypeMap,
         ClusterResponse,
         ReturnTypeAttribute,
+        ReturnTypeJson,
+        UniversalReturnTypeJson,
     };
 
     globalObject = Object.assign(global, nativeBinding);